Map companytype

+++ code/database/migrations/2022_09_20_154853_create_companies_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('companies', function (Blueprint $table) {
             $table->id();
-            $table->integer('type_id');
+            $table->unsignedBigInteger('type_id');
             $table->string('name');
             $table->timestamps();
         });
